-Remove DbgHelp types moved to ChaosLib -Fix EnumerableLocalProxyStub crashing when no notifier is specified -Fix MethodInfoDetailed.SigMethod crashing when a method is contained in a dynamic module -Add Get-ClrType cmdlet. Searches for types that match a specified predicate in a file or folder full of assemblies -Fix Appveyor build

+++ src/DebugTools/DbgHelp/SymbolManager.cs
@@ -50,7 +50,7 @@ namespace DebugTools
                 LoadModuleSymbols(module);
             }
 
-            var hr = DbgHelp.TrySymFromAddr(Process.Handle, address, out var result);
+            var hr = DbgHelp.TrySymFromAddr(Process.Handle, (long) address, out var result);
 
             if (hr != HRESULT.S_OK)
             {
